implement mixed precision

+++ SynthSeg/metrics_model.py
@@ -91,7 +91,7 @@ class WeightedL2Loss:
         return config
 
     def __call__(self, gt, pred):
-        gt = tf.one_hot(gt, depth=self._n_labels, axis=-1, dtype=tf.float32)
+        gt = tf.one_hot(gt, depth=self._n_labels, axis=-1, dtype=pred.dtype)
         weights = tf.expand_dims(1 - gt[..., 0] + self._background_weight, -1)
         return tf.keras.backend.sum(
             weights * tf.keras.backend.square(pred - self._target_value * (2 * gt - 1))
@@ -113,7 +113,7 @@ class DiceLoss:
         return config
 
     def __call__(self, gt, pred):
-        gt = tf.one_hot(gt, depth=self._n_labels, axis=-1, dtype=tf.float32)
+        gt = tf.one_hot(gt, depth=self._n_labels, axis=-1, dtype=pred.dtype)
         # make sure tensors are probabilistic
         if (
             self._enable_checks
